# docs/swagger.yaml
definitions:
...
  entity.Users:
    properties:
      address:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      platNo:
        type: string
      profilePicture:
        type: string
      updatedAt:
        type: string
      userName:
        type: string
    type: object
  entity.UsersAddress:
    properties:
      address:
        type: string
      address_name:
        type: string
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      person:
        type: string
      phone:
        type: string
      primary:
        type: boolean
      user_id:
        type: integer
    type: object
  entity.UsersConfig:
    properties:
      radius:
        type: integer
    type: object
...
  model.ReqChangePassword:
    properties:
      password:
        type: string
    required:
    - password
    type: object
...
  model.ReqResetPassword:
    properties:
      userName:
        type: string
    required:
    - userName
    type: object
...
  model.ReqUserCreate:
    properties:
      roleId:
        type: integer
      userName:
        type: string
    required:
    - roleId
    - userName
    type: object
  model.ReqUserLogin:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  model.ReqUserLoginGoogle:
    properties:
      idToken:
        type: string
      name:
        type: string
      phone:
        type: string
      profilePicture:
        type: string
    required:
    - idToken
    type: object
  model.ReqUserUpdate:
    properties:
      address:
        type: string
      name:
        type: string
      phone:
        type: string
      platNo:
        type: string
    type: object
...
  model.ResUserAll:
    properties:
      address:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      platNo:
        type: string
      profilePicture:
        type: string
      roleId:
        type: integer
      roleName:
        type: string
      updatedAt:
        type: string
      userName:
        type: string
    type: object
  model.ResUserMerchant:
    properties:
      merchantId:
        type: integer
      merchantName:
        type: string
      username:
        type: string
      usersId:
        type: integer
    type: object
...
  model.ResponseErrors:
    properties:
      error:
        $ref: '#/definitions/model.errorResponse'
    type: object
  model.ResponseSuccess:
    properties:
      data:
        type: object
    type: object
...
  model.errorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /check:
    get:
      operationId: CheckToken
      parameters:
      - description: token reset
        in: query
        name: token
        required: true
        type: string
      summary: Check Token Forget Password
      tags:
      - Authorization
...
  /forgotpassword:
    post:
      operationId: ForgotPassword
      parameters:
      - description: username
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.ReqResetPassword'
      responses:
        "200":
          description: 'message: "Send Email Forgot Password Success" '
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "422":
          description: 'code: 422, message: "Invalid request" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "500":
          description: 'code: 500, message: "Username not found" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      summary: Send Email Forgot Password
      tags:
      - Authorization
...
  /login:
    post:
      operationId: Login
      parameters:
      - default: streetbox-mobile-merchant
        description: ' '
        in: header
        name: CLIENT_ID
        required: true
        type: string
      - description: all fields mandatory
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/model.ReqUserLogin'
      responses:
        "200":
          description: 'token: "exampletokenresponse" '
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "401":
          description: 'code: 401, message: "username or password not valid, please
            try again" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "422":
          description: 'code: 422, message: "Invalid request" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      summary: Login user
      tags:
      - Authorization
  /login/google:
    post:
      operationId: LoginGoogle
      parameters:
      - description: Google
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/model.ReqUserLoginGoogle'
      responses:
        "200":
          description: 'message: "exampletokenresponse" '
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "401":
          description: 'code: 401, message: "username or password not valid, please
            try again" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "422":
          description: 'code: 422, message: "Invalid request" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      summary: Login End User Using Google
      tags:
      - Authorization
...
  /resetpassword:
    put:
      operationId: ResetPassword
      parameters:
      - description: change password user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.ReqChangePassword'
      responses:
        "200":
          description: 'message: "Reset Password Success" '
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "422":
          description: 'code: 422, message: "Invalid request" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "500":
          description: 'code: 500, message: "Reset Password Failed" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      summary: Reset password user (permission  = all)
      tags:
      - Authorization
...
  /user:
    get:
      operationId: GetAllUser
      parameters:
      - description: ' '
        in: query
        name: filter
        type: string
      responses:
        "200":
          description: 'data: []entity.Users'
          schema:
            items:
              $ref: '#/definitions/entity.Users'
            type: array
      security:
      - Token: []
      summary: Get All user (permission = superadmin)
      tags:
      - Master User
    post:
      operationId: CreateUser
      parameters:
      - description: new user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.ReqUserCreate'
      responses:
        "201":
          description: 'data: entity.Users, message: "Create User Success" '
          schema:
            $ref: '#/definitions/entity.Users'
        "422":
          description: 'code: 422, message: "Invalid request" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "500":
          description: 'code: 500, message: "Create User Failed" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Create User Superadmin or Admin (permission = superadmin)
      tags:
      - Master User
  /user/:id/address:
    delete:
      operationId: DeleteUserAddress
      parameters:
      - description: User address' id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{"message": "Success"}'
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "500":
          description: Delete user address failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Delete User Address (permission = consumer)
      tags:
      - Master User
  /user/{id}/delete:
    delete:
      operationId: DeleteByIDUser
      parameters:
      - description: userID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 'message: "Success'
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "500":
          description: 'code: 500, message: "Failed Delete User" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Delete Parking Space by ID (permission = superadmin)
      tags:
      - Master User
  /user/address:
    get:
      operationId: GetAddressByUserID
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            items:
              $ref: '#/definitions/entity.UsersAddress'
            type: array
        "500":
          description: Inquiry User Address Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Get User' Addresses (permission = consumer)
      tags:
      - Master User
    post:
      operationId: CreateAddress
      parameters:
      - description: id and primary fields are not mandatory
        in: body
        name: useraddress
        required: true
        schema:
          $ref: '#/definitions/entity.UsersAddress'
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            $ref: '#/definitions/entity.UsersAddress'
        "422":
          description: Invalid request
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "500":
          description: Create User Address Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Create User Address (permission = consumer)
      tags:
      - Master User
    put:
      operationId: UpdateUserAddress
      parameters:
      - description: all fields are not mandatory
        in: body
        name: useraddress
        required: true
        schema:
          $ref: '#/definitions/entity.UsersAddress'
      responses:
        "200":
          description: '{"message": "Success"}'
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "401":
          description: Invalid authorization
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "422":
          description: Invalid request
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "500":
          description: Create User Address Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Update User Address (permission = consumer)
      tags:
      - Master User
  /user/address/:id/switch:
    put:
      operationId: SwitchUserAddress
      parameters:
      - description: User address' id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{"message": "Success"}'
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "500":
          description: Switch Primary user address failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Switch Primary User Address (permission = consumer)
      tags:
      - Master User
  /user/address/primary:
    get:
      operationId: GetPrimaryAddressByUserID
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            $ref: '#/definitions/entity.UsersAddress'
        "500":
          description: Inquiry User Address Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Get User' Primary Address (permission = consumer)
      tags:
      - Master User
  /user/changepassword:
    put:
      operationId: ChangePassword
      parameters:
      - description: change password user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.ReqChangePassword'
      responses:
        "200":
          description: 'message: "Change Password Success" '
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "422":
          description: 'code: 422, message: "Invalid request" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "500":
          description: 'code: 500, message: "Change Password User Failed" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Change password user (permission = all)
      tags:
      - Master User
  /user/config:
    get:
      operationId: GetUserConfiguration
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            $ref: '#/definitions/entity.UsersConfig'
        "500":
          description: Inquiry User Configuration Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Get User' Configuration (permission = consumer)
      tags:
      - Master User
  /user/config/:rad:
    put:
      operationId: UpdateUserRadius
      parameters:
      - description: Radius in km
        in: path
        name: rad
        required: true
        type: string
      responses:
        "200":
          description: '{"message": "Success"}'
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "500":
          description: Update Users Radius Configuration Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Update User Radius (permission = superadmin)
      tags:
      - Master User
  /user/info:
    get:
      operationId: GetUser
      responses:
        "200":
          description: data:model.ResUserAll
          schema:
            $ref: '#/definitions/model.ResUserAll'
      security:
      - Token: []
      summary: Get user profile (permission = all)
      tags:
      - Master User
  /user/info/{id}:
    get:
      operationId: GetUserByID
      parameters:
      - description: userID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 'data: entity.Users'
          schema:
            $ref: '#/definitions/entity.Users'
      security:
      - Token: []
      summary: Get user profile by ID (permission = superadmin)
      tags:
      - Master User
  /user/merchant:
    get:
      operationId: GetUserMerchant
      responses:
        "200":
          description: 'data: []model.ResUserMerchant'
          schema:
            items:
              $ref: '#/definitions/model.ResUserMerchant'
            type: array
      security:
      - Token: []
      summary: Get All User Merchant Admin (permission = superadmin)
      tags:
      - Master User
  /user/role/{id}/update:
    put:
      consumes:
      - multipart/form-data
      operationId: UpdateUserRole
      parameters:
      - description: usersID
        in: path
        name: id
        required: true
        type: integer
      - description: Role ID
        in: formData
        name: roleID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: "Success" '
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "500":
          description: 'code: 500, message: "Failed Update Role User" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Update user role (permission = all)
      tags:
      - Master User
  /user/update:
    put:
      operationId: UpdateUser
      parameters:
      - description: update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.ReqUserUpdate'
      responses:
        "200":
          description: 'data: entity.Users'
          schema:
            $ref: '#/definitions/entity.Users'
        "422":
          description: 'code: 422, message: "Invalid request" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "500":
          description: 'code: 500, message: "Update User Failed" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Update user (permission = all)
      tags:
      - Master User
swagger: "2.0"